requirements.txt
flask==2.2.5
gunicorn==20.1.0
azure-monitor-opentelemetry
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
